+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Improvements to the Flow definition file. [@orta][]
 * Improve path generator for danger-runner.
 * Update the PR DSL to include bots. [@orta][]
+* Don't check for same user ID on comment when running as a GitHub App. [@tibdex][]
 
 ## 3.1.7